rename openvair installer stuff
