+++ rust-installer/src/openvair_manager/node_exporter.rs
@@ -1,1 +1,2 @@
+///! Модуль взаимодействия с утилитой node_exporter
 pub mod installer;
