+++ rust-installer/src/openvair_manager/pkg_management/distro/ubuntu.rs
@@ -36,5 +36,24 @@ impl PackageProvider for UbuntuPackageProvider {
     }
 }
 
+#[cfg(feature = "ubuntu")]
 #[cfg(test)]
-mod tests;
+mod tests {
+    use super::*;
+    use crate::tests::fixtures::command_runner;
+
+    #[rstest]
+    fn test_ubuntu_package_install(command_runner: CommandRunner) {
+        let provider = UbuntuPackageProvider::new(&command_runner);
+
+        let res = provider.try_install("hello");
+        assert!(
+            res.expect("failed to run install").status.success(),
+            "process did not terminate successfully"
+        );
+        assert!(
+            provider.check_installed("hello"),
+            "failed to find installed hello package"
+        );
+    }
+}
